requirements.txt
Flask==3.0.0
Flask_SocketIO
python-dotenv
confluent_kafka
protobuf